Significant progress on refactor + First calls being made for ELB creation
Continued my progress on the refactoring to a more generic workflow / config based engine. It's coming along nicely so far.

Additionally, I've plugged in the first pass on ELB creation. As of right now, I can create a launch config and the core of an ELB all from a single API call with config data loaded.