"Let it Crash"
Tuned up the supervisory strategy and now the system supports loosing connection with AWS in the middle of a workflow, it will restart the actor that failed to connect, and resend the message to try again. This process of failure recovery can happen for up to 5 minutes before it just gives up, it tries 15 seconds after a failure.