I'm not fully satisfied with this solution - but - it does follow closely with the way the Akka documentation recommends creating a factory to build your actor refs when you need to test parents in isolation. If I wanted to make it better (to get back the compile-time type-safety of actor creation), currently the only way I can think of is have a factory trait per actor's companion object. Then, have a provider that builds the factories... a lot of boilerplate.
Regardless - In the way I did this, I should be able to use the same ActorFactory for the entire production stack, and then when testing, I can have a test actor factory that matches only on the types of actors I want to mock or replace with probes, and then send all other matches back to the "production" stack.  So in this manner, I can test an entire actor hierarchy where I might only want to stub the very "bottom" actors that fetch to external systems while allowing all other parts of the system to operate like normal. We'll see how this goes as I start "hakking" it out.