fixed issue with existing stack + version.
This is a bug fix for the scenario in which you try to deploy to an existing stack with the exact same version. Normally, a new version would kick off a blue / green deploy. However, by convention the version number is in the stack name so a deploy with the same version will cause cloud formation to thrown an error - as we would want, however, we weren't handling this scenario. It ended up kicking the whole process into a fail state which made Chadash believe that it needed to tear down this "new stack" so it could unfreeze the old one. In this scenario, it actually ended up deleting the only running stack.

This issue should now be fixed and has a test to check for the scenario. A Deployment Failed status will be applied to any deploy request if the stack + version already exist.