Use AWSCredentialsProvider instead of passing around AWSCredentials
The InstanceProfileCredentials provider in the SDK knows when to automatically refresh to get the next credential set when getCredentials is called. The whole idea of having akka do this for us with a 30 minute refresh was a bad idea - leaving the alone for now, but it should be removed in the future as it's not needed. However, it is not hurting anything.

By passing the CredentialsProvider around, the correct / refreshed credentials should be provided automatically by AWS's SDK every time.